Remove RubyMotion trail
My goal with this commit is to present a more opinionated trail map for
students. thoughtbot currently writes our iOS applications in
Objective-C. That's where our expertise now resides, and is where we can
provide the most guidance. I think it's important to keep the trails
focused and not over-burden learners with choices. Our job is primarily
as curators. So, I'm pruning this trail out to focus the Learn offering
and reduce our maintenance and support burden.